+++ frontend/src/metabase/common/components/MetadataInfo/ColumnFingerprintInfo/QueryColumnFingerprintInfo.tsx
