+++ frontend/src/metabase/common/components/MetadataInfo/ColumnFingerprintInfo/index.ts
@@ -1,1 +1,1 @@
-export * from "./ColumnFingerprintInfo";
+export * from "./QueryColumnFingerprintInfo";
