+++ frontend/src/metabase/common/components/MetadataInfo/ColumnInfo/QueryColumnInfo.styled.tsx
