+++ frontend/src/metabase/common/components/MetadataInfo/ColumnInfoIcon/index.tsx
@@ -1,1 +1,1 @@
-export * from "./ColumnInfoIcon";
+export * from "./QueryColumnInfoIcon";
